crud color + loai os sua

+++ src/main/java/com/example/world_phone/dto/respone/attribute/screen/ScreenReposne.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class ScreenReposne {
     private Long id;
     private String name;
-    private int loai;
+    private Long loai;
 }
